ViaDelivers/VTP/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alabaster==0.7.10
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.22.0        # via cryptography
babel==2.4.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==2.0.0
boto3==1.4.4
boto==2.46.1
botocore==1.5.47
celery==3.1.25
cffi==1.10.0
chardet==3.0.2
collectfast==0.5.2
colorama==0.3.9
cryptography==1.8.1
cssselect==1.0.1
defusedxml==0.5.0         # via python3-openid, social-auth-core
django-after-response==0.2.2
django-appconf==1.0.2
django-autoslug==1.9.3
django-bootstrap-form==3.2.1
django-bootstrap-pagination==1.6.2
django-breadcrumbs==1.1.3
django-celery==3.2.1
django-configurations==2.0
django-dbindexer==1.6.1
django-extensions==1.7.9
django-filter==1.0.2
django-grappelli==2.9.1
django-heroku-memcacheify==1.0.0
django-impersonate==1.1
django-ipware==1.1.6
django-mathfilters==0.4.0
django-menus==1.1.2
django-nullablecharfield==0.2.2
django-nvd3==0.9.7
django-pylibmc==0.6.1     # via django-heroku-memcacheify
django-salesforce==0.7.2
django-storages==1.5.2
django-tables2==1.6.1
django-tagging==0.4.5
django-tinymce4-lite==1.4.0
django-widget-tweaks==1.4.1
django==1.10.8
djangorestframework==3.6.2
docutils==0.13.1
drfdocs==0.0.11
ecdsa==0.13
enum34==1.1.6
filebrowser-safe==0.4.7
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
grappelli-safe==0.4.6
html5lib==0.999999999
idna==2.5
imagesize==0.7.1
ipaddress==1.0.18
jinja2==2.9.6
jmespath==0.9.2
kombu==3.0.37
librabbitmq==1.6.1
lxml==3.7.3
markdown==2.6.8
markupsafe==1.0
mock==2.0.0
newrelic==2.86.0.65
oauthlib==2.0.2
olefile==0.44             # via pillow
packaging==16.8           # via cryptography
paramiko==2.1.2
pbr==3.0.0
pillow==4.1.1
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pygeoip==0.3.2
pygments==2.2.0
pyjwt==1.5.0
pylibmc==1.5.2
pyparsing==2.2.0
pyquery==1.2.17
python-dateutil==2.6.0
python-docx==0.8.6
python-memcached==1.58
python-nvd3==0.14.2
python-openid==2.2.5
python-slugify==1.1.4
python-social-auth==0.3.6
python3-openid==3.1.0     # via social-auth-core
pytz==2017.2
pyyaml==3.12
requests-oauthlib==0.8.0
requests==2.14.2
s3transfer==0.1.10        # via boto3
setuptools-git==1.2
simplejson==3.10.0
six==1.11.0
snowballstemmer==1.2.1
social-auth-core==1.3.0   # via python-social-auth
sphinx-rtd-theme==0.2.4
sphinx==1.5.5
sqlparse==0.2.3
tzlocal==1.4
unicodecsv==0.14.1
unidecode==0.4.20
waitress==1.0.2
webencodings==0.5.1       # via html5lib
wsgiref==0.1.2
